+++ clj.java/test/org/soulspace/clj/java/test/reflection.clj
@@ -1,7 +1,0 @@
-(ns org.soulspace.clj.java.test.reflection
-  (:use
-    [clojure.test]
-    [org.soulspace.clj.java.reflection]))
-
-(deftest field-seq-test
-  (is (= true true)))