+++ clj.java/test/org/soulspace/clj/java/test/type_conversion.clj
@@ -1,7 +1,0 @@
-(ns org.soulspace.clj.java.test.type-conversion
-  (:use
-    [clojure.test]
-    [org.soulspace.clj.java.type-conversion]))
-
-(deftest get-method-test
-  (is (= true true)))